feat: move req and resp to application and fix test command
