+++ src/contexts/core/genially/application/models/GeniallyRequest.ts
