+++ src/contexts/core/genially/application/models/GeniallyResponse.ts
